+++ src/_pytask/click.py
@@ -242,11 +242,11 @@ def _format_help_text(  # noqa: C901, PLR0912, PLR0915
 
     if show_default_is_str or (show_default and (default_value is not None)):
         if show_default_is_str:
-            default_string = f"({param.show_default})"  # type: ignore[attr-defined]
+            default_string = param.show_default  # type: ignore[attr-defined]
         elif isinstance(default_value, (list, tuple)):
             default_string = ", ".join(str(d) for d in default_value)
         elif inspect.isfunction(default_value):
-            default_string = _("(dynamic)")
+            default_string = _("dynamic")
         elif param.is_bool_flag and param.secondary_opts:  # type: ignore[attr-defined]
             # For boolean flags that have distinct True/False opts,
             # use the opt without prefix instead of the value.