+++ src/pytask/__init__.py
@@ -13,7 +13,9 @@ from _pytask.compat import check_for_optional_program
 from _pytask.compat import import_optional_dependency
 from _pytask.config import hookimpl
 from _pytask.console import console
-from _pytask.database_utils import db
+from _pytask.database_utils import create_database
+from _pytask.database_utils import DatabaseSession
+from _pytask.database_utils import State
 from _pytask.exceptions import CollectionError
 from _pytask.exceptions import ConfigurationError
 from _pytask.exceptions import ExecutionError
@@ -44,6 +46,7 @@ from _pytask.outcomes import Skipped
 from _pytask.outcomes import SkippedAncestorFailed
 from _pytask.outcomes import SkippedUnchanged
 from _pytask.outcomes import TaskOutcome
+from _pytask.profile import Runtime
 from _pytask.report import CollectionReport
 from _pytask.report import DagReport
 from _pytask.report import ExecutionReport
@@ -69,6 +72,8 @@ __all__ = [
     "ColoredCommand",
     "ColoredGroup",
     "ConfigurationError",
+    "DagReport",
+    "DatabaseSession",
     "EnumChoice",
     "ExecutionError",
     "ExecutionReport",
@@ -84,11 +89,12 @@ __all__ = [
     "Persisted",
     "PytaskError",
     "ResolvingDependenciesError",
-    "DagReport",
+    "Runtime",
     "Session",
     "Skipped",
     "SkippedAncestorFailed",
     "SkippedUnchanged",
+    "State",
     "Task",
     "TaskOutcome",
     "WarningReport",
@@ -98,7 +104,7 @@ __all__ = [
     "cli",
     "console",
     "count_outcomes",
-    "db",
+    "create_database",
     "depends_on",
     "format_exception_without_traceback",
     "get_all_marks",