+++ docs/source/changes.md
@@ -8,6 +8,7 @@ releases are available on [PyPI](https://pypi.org/project/pytask) and
 ## 0.4.0 - 2023-xx-xx
 
 - {pull}`323` remove Python 3.7 support and use a new Github action to provide mamba.
+- {pull}`387` replaces pony with sqlalchemy.
 
 ## 0.3.2 - 2023-06-07